Up the range of fibonacci by using u128

+++ projects/fibonacci/src/main.rs
@@ -18,7 +18,7 @@ fn main() {
     }
 }
 
-fn get_nth_fibonacci(desired_index: u32) -> usize {
+fn get_nth_fibonacci(desired_index: u32) -> u128 {
     let mut current_index = 0;
 
     let mut preceding = 0;
